Extract all btree tests into a reusable test suite
